service-bus-to-delta-lake-func/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
databricks-connect==11.3.0b0
jsonpath_ng
xmltodict